Add files via upload
In this thesis, I studied the topological phase transition induced by external small perturbations beyond the general “Floquet engineering”. We constructed and focused on the “SSH-Holstein” model, then we mapped the band topology to the classical magnetization, and I derived an analog equation of the Landau-Lifshitz-Gilbert equation to describe the motion of the order parameter. Moreover, I found that this equation is deeply connected to the chiral anomaly. Consequently, I proved that the two different topological states could form a “topological bit”, as an analogue of the magnetic bit.